requirements.txt
django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-simple-history==3.10.1
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.4.5
qrcode==8.2
pillow==11.0.0
